This commit sets commitFlag, noDiskFlag, startFlag and executeFlag correctly for linked operations. (Previously, they had hard-coded values.)
New flags has been added to DABits::NodeInfoBits and DABits::ParamInfoBits to specify wether disk columns are 
present in the user projection or in the projection that the SPJ block uses to instantiate child operations. These 
flags are used by the SPJ block for setting noDiskFlag properly in the ScanFragReq and LqhKeyReq signals that 
it sends to the LQH.

This commit also fixes an error concerning life time of NdbQueryDef objects: Prevuously, the ndb handler would 
release an NdbQueryDef object as soon as the query was complete. But the transaction would keep the corresponding 
NdbQuery objects until the transaction ended. Since an NdbQueryDef object is required to live longer than any 
NdbQuery object derived from it, this was an error. The ndb handler will now keep a list of NdbQueryDef objects per 
transaction, and release these when the transaction ends.