Bug#51896  :make test-bt terminates too early with cmake.
Reason for it is test-bt is a custom target executing  bunch of 
COMMAND(perl mysql-test-run.pl ...). If one of the COMMAND fails, the
whole build termintates. 
      
The fix is to write a cmake script run_collection that takes a
corresponding "collections" file and executes it line by line with
EXECUTE_PROCESS and ignore all EXECUTE_PROCESS errors.
      
- Also, added or fixed copyright notices.