Bug#52134 performance schema file io, symlink in path
Before this fix, a file named "/path/with/link/to/filename"
would be normalized to:
- "/path/with/link/to/filename" when the file did not exist (creation),
- "/real/path/to/filename" when the file did exist (file io, deletion),
causing instrumentation leak for file io in the performance schema.

With this fix, file name normalization has been revisited to be more robust,
and always resolve this case to "/real/path/to/filename".

File instrumentation now works properly, even when a symbolic link
is present in the directory name, which is the case when running tests
with MTR_MEM=1, where the var directory is a link to shared memory.