Bug#44303 Assertion failures in Field_new_decimal::store_decimal when executing           materialized InsideOut semijoin
(Backporting of revid:oystein.grovlen@sun.com-20091009093541-7bkrgk2tsnpi6iyf)

When inner tables of semijoin is materialized, field values are copied back
from the materialized table to the record buffer of the source tables.
The problem was that the write_set of the source tables were not set up to
allow such copying.