Bug#46692 Crash occurring on queries with nested FROM subqueries           using materialization
(Backporting of revid:oystein.grovlen@sun.com-20091229132523-qv9gud3wyrb1ps4a)

Subqueries with STRAIGHT_JOIN prevents pull-out of tables that would
normally be pulled out of the subquery.  This created scenarioes which
the plan search was not prepared for, and one ended up without any valid
plan.  This caused segmentation fault in optimize_semijoin_nests when
trying to access the best plan.

The problem may occur with either const tables or eq_ref relations
that does not have any matches.  Bug#46692 reports the problem for
FROM subqueries, but ordinary IN subqueries will have the same issue.

The following changes are needed to handle this problem:
1. Do not include const tables in the set of tables to be considered
   by optimize_semijon_nests().  This will make sure cost estimation is
   not skipped for tables that depend on const tables.  (This is
   consistent with what is done for top level joins where const tables
   are filtered out from the tables considered by choose_plan.)

2. Change computation of record count so that division by inner fanout
   is not needed.  When the table that is added to the join prefix, is
   empty, fanout will become zero since it is the product of the
   number of records of all tables.  When using division, a zero inner
   fanout caused the computed cost to become infinite.

The fanout issue (issue 2 above) is also present with the duplicate
weedout strategy, and this patch contains the necessary changes to
avoid division in this scenario, too.

A number of bug reports have been marked as duplicate of this bug report
based on that this fix also fixes these issues.  Those bug reports show
that this is not just an issue for STRAIGHT_JOIN, but also for both
right and left outer join.  Note that while this fix, fixes the issue
for outer joins, a more proper fix would be to make sure that the
dependent constant tables could be pulled out.  This will be fixed by
bug 49952.

This patch also adds some test cases for these duplicate reports as well
as general testing of outer joins in subqueries when constant tables are
involved.