BUG#44209: MASTER_CONNECT_RETRY and --master-retry-count disconnected from each other
MASTER_CONNECT_RETRY and --master-retry-count are logically
related. The former sets the value for the time the slave thread
waits until reconnecting, while the latter sets the value for 
the number of reconnection attempts until it gives up. 
  
The problem here was that --master-retry-count is not dynamic
and can only be set on slave startup, while MASTER_CONNECT_RETRY
is dynamic and can be set using CHANGE MASTER. As such, the user
should also be allowed to change the value of master_retry_count
using CHANGE MASTER command.
  
In this patch we extend the syntax of CHANGE MASTER to allow 
MASTER_RETRY_COUNT symbol that can be used to set the variable
that holds the value for master_retry_count. The option 
--master-retry-count can now be deprecated (this patch takes no
action regarding deprecating this option).