BUG#45863: Assertion failed: (fixed == 0)...
The problem in JOIN::prepare() is that fix_fields() is called prematurely
from subquery_types_allow_materialization().
The call to fix_fields() observed the unknown column name and set error
status accordingly.
But JOIN::prepare() failed to notice the error and attempted to call
fix_fields() again, now with a half-way resolved expression, and we have
an assert().
The fix is mostly taken from the committed but unpushed work in WL#4389,
which is about applying semijoin transformation to EXISTS queries,
but also consolidates some of the semantic checking on subqueries.
The call to fix_fields() inside JOIN::prepare() is moved so that it is
always executed before subquery_types_allow_materialization().
Hence, the latter no longer needs to call fix_fields() and we can remove
the thd argument (and also let both return values become output arguments).
I also removed one of the "psergey-todo" comments, because the subselect
cardinality check is now always performed in the early stages of
JOIN::prepare(). But eliminating all the comments would be beyond the scope
of this bugfix.
Noticed another problem during development as well: One query was optimized
differently in regular and prepared mode, and this happened because the
is_correlated flag was set differently in the two modes. Had to make sure
that this flag is set consistently, by calling fix_fields() from
JOIN::prepare() also for the first execution of a prepared statement.
Otherwise, is_correlated would be set inconsistently by the fix_fields()
call in convert_subq_to_sj().
  
mysql-test/r/subselect4.result
  Added test results for bug.
        
mysql-test/t/subselect4.test
  Added tests for bug.
        
sql/sql_select.cc
  Modified JOIN::prepare() and subquery_types_allow_materialization()
  according to above description.