Bug#48093: 6.0 Server not processing equivalent IN clauses            properly with Innodb tables
(Backporting of jorgen.loland@sun.com-20091106074019-cqoxr3hvmbgqcldt)
      
When MRR was used to lookup multiple values in an index that
return ROR ordered records, it was assumed that the returned
records from MRR were also ROR ordered. This is not true since
MRR returns all records retrieved for the first value, then all
records for the second value and so on. Although the records are
ROR ordered for each value, the end result is not ROR order.
      
Since the optimizer thought multi-value lookups returned ROR
ordered records, it tried to do index merge on the returned
records. These turned out not to be ROR ordered, resulting in
incorrect merging and in turn missing rows in the result set.
      
The fix is to invalidate ROR ordering for index lookup 
algorithms when looking up multiple key values.