Fix for Bug#36981 innodb crash when selecting for update.
(Backporting of revid:olav@sun.com-20100225125531-9249z4xxvxqya96u)
                  
This crash occurred due to a mismatch between what the ICP code in
InnoDB expected to be in the InnoDB record and what actually was there. 
The ICP code expected the record to contain an index entry for the 
index for the ICP condition while the record contained an entry from
the clustered index. This happened due to InnoDB changing from using
the ICP index to using the clustered index if the statement should use
LOCK_X type locks (which happens when doing SELECT...FOR UPDATE or
DELETE operations).
    
The ICP implementation in InnoDB does not support ICP if the index is
the primary key (using the clustered index). This change fixes the
crash by extending the check for whether or not to accept the pushed
index to also not support the case where InnoDB will change from using
the ICP index to the clustered index when the lock type is LOCK_X.
                  
The attached test case contains two tests. The first is based on the
original reproduction case from the bug. The second test case is added
to have a test to verify that the condition is evaluated even if it
is refused by InnoDB.