BUG#49906: Assertion failed - Field_varstring::val_str in field.cc
(Backporting of revid:jorgen.loland@sun.com-20100126063016-f4vyz4jcymk2urmt)

When filesort reads sort keys, fields that are evaluated by pushed 
conditions were not in the read_set. This resulted in an ASSERT
when the engine evaluated the pushed conditions.

The fix is to add fields evaluated by pushed conditions to the 
read_set.