Fix for Bug#50381 Assertion failing in handler.h:1283: void COST_VECT::add_io(double, double)
(backporting of revid:olav.sandstaa@sun.com-20100316175401-iax4ockzqybjrjpl)
      
The crash occured when an assert was checking that the value for the
average IO cost in the cost vector did not have the value NaN. The NaN
value was introduced in the cost vector object in get_sweep_read_cost()
when the size of the data file for the table was 0 byte (ie. empty
data file).  This resulted in a division by zero when computing the
estimated number of blocks needed to be read from disk.
      
The fix for avoiding the division by zero is to assume that the data
file always contains at least one block. This is similar to what is
done for the estimated number of blocks to be accessed. So after this
fix the cost estimate in get_sweep_read_cost() will assume both that
the data file and the estimate for number of blocks to access are at
least 1.