Bug#51084: "Batched key access crashes for SELECT with derived             table and LEFT JOIN"
(Backporting of revid:jorgen.loland@sun.com-20100317122515-qrhfyp5xhqy07wr9)

JOIN::optimize() may decide to use a join cache to join a table,
but may have to revise this decision later, e.g. if an 
incompatible access method is chosen. If it is decided that a
join cache cannot be employed after all, set_join_cache_denial()
is called to destroy the cache object. However, if linked join
buffers are in use, a join cache for another table may link to 
the cache being removed through the next_cache pointer. The 
crash happened when the destroyed object was tried accessed 
through this link.

The fix is to remove the next_cache pointer from the previous 
cache if the cache being removed has a previous cache.