Bug#49952 - Inner tables in subquery outer join is not detected as constant tables
(Backporting of oystein.grovlen@sun.com-20100417063402-gmwfvs9mmu2u714k)

Problem: If a subquery that has been converted to semijoin contains 
an outer join, constant table optimization is not always done.  
This happens because the necessary test to skip constant table 
optimization in case of nested outer join gives false hits for 
semijoin nests.

Fix: Introduce a new function TABLE_LIST::in_outer_join_nest() that correctly
identifies when the table belongs to a nested join, and use this function
where special handling of nested joins are required.  Also replaced correct
tests for nested join with this function. 

Also, fixed a check for dependence on const tables that was not
based on the set of all const tables, but just the set of tables with
one row.  Hence, it was not always detected that constant table
optimization could be used for inner tables of outer joins when the 
outer table was empty.