Bug#42580: "Innodb's ORDER BY ..LIMIT returns no rows for             null-safe operator <=> NULL"
When Index Condition Pushdown is used in InnoDB, the row returned
to MySQL contains both the index columns and the table columns.
Before, the null bit of the index columns were not set correctly 
in the MySQL formated record. This resulted in missing records 
in queries using the <=> operator.

This patch also fixes bugs 43617 and 43249. Regression tests for
these bugs are included.
     
The patch was originally submitted by Igor and has been through 
SCA.