Modified SPJ API to also accept (constant) scanfilters for child operations:
Mainly relax the checking of which operation type we have in 
NdbQueryOperationImpl::setInterpretedCode. From now on ::setInterpretedCode()
should be possible on any operation type - Removed setErrorCodeAbort(QRY_WRONG_OPERATION_TYPE)
if the operation was not a ScanOperation.

Ensure that ::doSend() ::setInterpretedFlag() for a lookup root containing scanFilters.
In order to determine if an operation has a scanfilter, the method
 NdbQueryOperationImpl::hasInterpretedCode() has been added.

Also renamed NdbQueryOperationImpl::prepareScanFilter to 
NdbQueryOperationImpl::prepareInterpretedCode to be more consistent with names on
other methods related to interpretedCode usage.