BUG#44209: MASTER_CONNECT_RETRY and --master-retry-count disconnected from each other
Post-push fix.

The mi->retry_count was set to an uninitilized value when master.info
did not contain a correspondent line for the retry_count value. We fix
this by setting the default before actually loading the value from
master.info. If there is no value to read, mi->retry_count remains the
default (or the value set in the command line). The retry_count in 
master.info value takes precedence over the one from the command line
(which I believe is more inline with the proposal for deprecation of 
master-retry-count).