Patch for Bug #40154 main.subselect_sj2 fails sporadically on pushbuild
With Index Condition Pushdown (ICP) enabled one of the select queries
in subselect_sj2 failed randomly. As a "work-around" for this test
failure, ICP has earlier been disabled for this select query.

The cause for the test failure in subselect_sj2 has likely been fixed
by one of the earlier bug fixes related to ICP and InnoDB. This patch
removes the previously introduced disabling of ICP for this select
query.